+++ k8s-config.tf
@@ -1,3 +1,9 @@
+/**
+ * Copyright (C) 2019-2021 Expedia, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ */
+
 resource "kubernetes_secret" "beekeeper_db_password_secret" {
   metadata {
     name      = var.k8s_db_password_secret