+++ k8s.tf
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2019-2020 Expedia, Inc.
+ * Copyright (C) 2019-2021 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  */
@@ -18,6 +18,7 @@ resource "kubernetes_ingress" "beekeeper" {
     name        = local.k8s_app_alias
     labels      = local.labels
     annotations = {}
+    namespace = var.k8s_namespace
   }
 
   spec {
@@ -32,8 +33,8 @@ resource "kubernetes_ingress" "beekeeper" {
         path {
           path = var.k8s_path_cleanup_ingress_path
           backend {
-            service_name = kubernetes_service.beekeeper_path_cleanup[count.index].metadata.name
-            service_port = kubernetes_service.beekeeper_path_cleanup[count.index].spec.port.target_port
+            service_name = kubernetes_service.beekeeper_path_cleanup[0].metadata[0].name
+            service_port = kubernetes_service.beekeeper_path_cleanup[0].spec[0].port[0].target_port
           }
         }
       }
@@ -45,8 +46,8 @@ resource "kubernetes_ingress" "beekeeper" {
         path {
           path = var.k8s_metadata_cleanup_ingress_path
           backend {
-            service_name = kubernetes_service.beekeeper_metadata_cleanup[count.index].metadata.name
-            service_port = kubernetes_service.beekeeper_metadata_cleanup[count.index].spec.port.target_port
+            service_name = kubernetes_service.beekeeper_metadata_cleanup[0].metadata[0].name
+            service_port = kubernetes_service.beekeeper_metadata_cleanup[0].spec[0].port[0].target_port
           }
         }
       }
@@ -58,8 +59,8 @@ resource "kubernetes_ingress" "beekeeper" {
         path {
           path = var.k8s_scheduler_apiary_ingress_path
           backend {
-            service_name = kubernetes_service.beekeeper_scheduler_apiary[count.index].metadata.name
-            service_port = kubernetes_service.beekeeper_scheduler_apiary[count.index].spec.port.target_port
+            service_name = kubernetes_service.beekeeper_scheduler_apiary[0].metadata[0].name
+            service_port = kubernetes_service.beekeeper_scheduler_apiary[0].spec[0].port[0].target_port
           }
         }
       }