+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [3.2.1] - 2021-08-27
+### Added
+- Changed the service namespace from `default` to `beekeeper` in `beekeeper-metadata-cleanup`, `beekeeper-path-cleanup` and `beekeeper-scheduler-apiary`.
+- Changed some variables from type `string` to `number`.
+- Fixed the `beekeeper` ingress so that the `k8s_ingress_enabled` variable works.
+
 ## [3.2.0] - 2021-08-13
 ### Added
 - Terraform files for `beekeeper-api` module.